+++ src/apis/acl_entry_api.rs
@@ -126,7 +126,7 @@ pub enum UpdateAclEntryError {
 }
 
 
-/// Update multiple ACL entries on the same ACL.
+/// Update multiple ACL entries on the same ACL. For faster updates to your service, group your changes into large batches. The maximum batch size is 1000 entries. [Contact support](https://support.fastly.com/) to discuss raising this limit.
 pub async fn bulk_update_acl_entries(configuration: &mut configuration::Configuration, params: BulkUpdateAclEntriesParams) -> Result<crate::models::InlineResponse200, Error<BulkUpdateAclEntriesError>> {
     let local_var_configuration = configuration;