+++ src/models/inline_response_200_4.rs
@@ -9,16 +9,18 @@
 
 
 #[derive(Clone, Debug, PartialEq, Default, Serialize, Deserialize)]
-pub struct InlineResponse2002 {
-    /// Time-stamp (GMT) when the domain_ownership validation will expire.
-    #[serde(rename = "expires_at", skip_serializing_if = "Option::is_none")]
-    pub expires_at: Option<String>,
+pub struct InlineResponse2004 {
+    #[serde(rename = "data", skip_serializing_if = "Option::is_none")]
+    pub data: Option<Vec<String>>,
+    #[serde(rename = "meta", skip_serializing_if = "Option::is_none")]
+    pub meta: Option<Box<crate::models::InlineResponse2004Meta>>,
 }
 
-impl InlineResponse2002 {
-    pub fn new() -> InlineResponse2002 {
-        InlineResponse2002 {
-            expires_at: None,
+impl InlineResponse2004 {
+    pub fn new() -> InlineResponse2004 {
+        InlineResponse2004 {
+            data: None,
+            meta: None,
         }
     }
 }